+++ src/go/cmd/godplugin/main.go
@@ -15,9 +15,9 @@ import (
 	"github.com/netdata/netdata/go/plugins/logger"
 	"github.com/netdata/netdata/go/plugins/pkg/buildinfo"
 	"github.com/netdata/netdata/go/plugins/pkg/executable"
+	"github.com/netdata/netdata/go/plugins/pkg/multipath"
 	"github.com/netdata/netdata/go/plugins/plugin/go.d/agent"
 	"github.com/netdata/netdata/go/plugins/plugin/go.d/cli"
-	"github.com/netdata/netdata/go/plugins/plugin/go.d/pkg/multipath"
 
 	"github.com/jessevdk/go-flags"
 	"golang.org/x/net/http/httpproxy"
