+++ src/go/plugin/go.d/examples/simple/main.go
@@ -10,13 +10,13 @@ import (
 	"os"
 	"path"
 
+	"github.com/jessevdk/go-flags"
+
 	"github.com/netdata/netdata/go/plugins/logger"
+	"github.com/netdata/netdata/go/plugins/pkg/multipath"
 	"github.com/netdata/netdata/go/plugins/plugin/go.d/agent"
 	"github.com/netdata/netdata/go/plugins/plugin/go.d/agent/module"
 	"github.com/netdata/netdata/go/plugins/plugin/go.d/cli"
-	"github.com/netdata/netdata/go/plugins/plugin/go.d/pkg/multipath"
-
-	"github.com/jessevdk/go-flags"
 )
 
 var version = "v0.0.1-example"
